Backend initializer branch

+++ server/src/main/java/com/lankacapital/server/dtos/LoanCreateDto.java
@@ -10,9 +10,9 @@ import java.math.BigDecimal;
 public class LoanCreateDto {
     private String fileNumber;
     private BigDecimal loanAmount;
-    private Double interestRate;
+    private Integer interestRate;
     private BigDecimal documentCharge;
-    private Double numberOfInstallments;
+    private Integer numberOfInstallments;
     private Long employeeId;
 
     private Long customerId;
